backend/requirements.txt
anyio==3.6.2
APScheduler==3.6.3
asgiref==3.6.0
cachetools==4.2.2
certifi==2022.12.7
charset-normalizer==3.0.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.1
Django==4.1.5
django-cors-headers==3.13.0
django-environ==0.9.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.4
flake8==6.0.0
flake8-broken-line==1.0.0
flake8-isort==6.0.0
flake8-plugin-utils==1.3.2
flake8-return==1.2.0
gunicorn==20.0.4
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
inflection==0.5.1
isort==5.12.0
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
mccabe==0.7.0
packaging==23.0
pep8-naming==0.13.3
Pillow==10.4.0
postgres==4.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
psycopg2-pool==1.1
pyaes==1.6.1
pyasn1==0.5.0
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PyJWT==2.6.0
Pyrogram==2.0.102
PySocks==1.7.1
pyTelegramBotAPI==4.10.0
python-json-logger==2.0.7
python-telegram-bot==13.7
pycryptodome==3.18.0
python-jose==3.3.0
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.2
rfc3986==1.5.0
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.3
telebot==0.0.5
tornado==6.2
tzdata==2022.7
tzlocal==4.3
uritemplate==4.1.1
urllib3==1.26.14
whitenoise==6.3.0
